+++ kgsm-lib/Json/KgsmJsonContext.cs
@@ -45,6 +45,10 @@ namespace TheKrystalShip.KGSM;
 [JsonSerializable(typeof(WatchdogPlayer))]
 [JsonSerializable(typeof(WatchdogPlayer[]))]
 [JsonSerializable(typeof(Dictionary<string, WatchdogPlayer[]>))]
+[JsonSerializable(typeof(WatchdogUpnpMapping))]
+[JsonSerializable(typeof(WatchdogUpnpList))]
+[JsonSerializable(typeof(WatchdogUpnpActionResult))]
+[JsonSerializable(typeof(WatchdogUpnpOpenRequest))]
 // --- Event envelope + every event data payload (the Type-keyed dispatch in EventService) ---
 [JsonSerializable(typeof(EventWrapper))]
 [JsonSerializable(typeof(InstanceCreatedData))]